// src/api/models/MeasurementRetrieve.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MeasurementRetrieve = {
    readonly measurement_id: number;
    readonly created_at: string;
    readonly updated_at: string;
    measurement_date?: string | null;
    person?: number | null;
    measurement_concept?: number | null;
    measurement_type_concept?: number | null;
};